fix init job error
