+++ pkg/cli/initconfig/cmd/healthz.go
@@ -1,3 +1,18 @@
+/*
+Copyright 2021 The KodeRover Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
 package cmd
 
 import (
@@ -44,9 +59,9 @@ func (initCheck) Check(req *http.Request) error {
 }
 
 func checkUserServiceHealth() error {
-	return user.New().GetUserSvrHealthz()
+	return user.New().Healthz()
 }
 
 func checkPolicyServiceHealth() error {
-	return policy.NewDefault().GetPolicySvrHealthz()
+	return policy.NewDefault().Healthz()
 }
