+++ pkg/microservice/policy/core/service.go
@@ -108,3 +108,7 @@ type indexer interface {
 	EnsureIndex(ctx context.Context) error
 	GetCollectionName() string
 }
+
+func CheckMongodbHealth(ctx context.Context) error {
+	return mongotool.Ping(ctx)
+}
