+++ pkg/microservice/user/core/service.go
@@ -83,3 +83,17 @@ func initDatabase() {
 func Stop(_ context.Context) {
 	gormtool.Close()
 }
+
+func Healthz() error {
+	sqlDB, err := DB.DB()
+	if err != nil {
+		log.Errorf("Healthz get db error:%s", err.Error())
+		return err
+	}
+	err = sqlDB.Ping()
+	if err != nil {
+		log.Errorf("Healthz ping mysql db error:%s", err.Error())
+		return err
+	}
+	return nil
+}
