+++ pkg/microservice/user/core/service/user/user.go
@@ -563,13 +563,3 @@ func SyncUser(syncUserInfo *SyncUserInfo, logger *zap.SugaredLogger) (*models.Us
 	}
 	return user, nil
 }
-
-// UserSvrHealthz
-func UserSvrHealthz(logger *zap.SugaredLogger) error {
-	err := orm.CheckUserMysqlDbHealthz()
-	if err != nil {
-		logger.Errorf("CheckUserMysqlDbHealthz error,error msg:%s ", err)
-		return err
-	}
-	return nil
-}
